move EncodingOracle to the AbstractTextFileService for #79275
